Graduation_Projects/Project1_SAHLA/sign2speech/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
elevenlabs==1.50.2
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-texttospeech==2.22.0
googleapis-common-protos==1.66.0
gradio==5.9.1
gradio_client==1.5.2
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
jax==0.4.37
jaxlib==0.4.36
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf>=3.20,<6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
ruff==0.8.3
safehttpx==0.1.6
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.41.3
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
websockets==14.1
